TUP-28334 refactor
