+++ main/plugins/org.talend.commons.runtime/src/org/talend/commons/CommonsPlugin.java
@@ -15,15 +15,12 @@ package org.talend.commons;
 import java.io.IOException;
 import java.io.InputStream;
 import java.net.URL;
-import java.util.Properties;
 
 import org.apache.commons.lang.ArrayUtils;
 import org.eclipse.core.net.proxy.IProxyService;
 import org.eclipse.core.runtime.FileLocator;
-import org.eclipse.core.runtime.IStatus;
 import org.eclipse.core.runtime.Platform;
 import org.eclipse.core.runtime.Plugin;
-import org.eclipse.core.runtime.Status;
 import org.eclipse.core.runtime.preferences.InstanceScope;
 import org.osgi.framework.Bundle;
 import org.osgi.framework.BundleContext;
@@ -90,8 +87,6 @@ public class CommonsPlugin extends Plugin {
 
     private BundleContext context;
 
-    private PerformanceLogManager logManager;
-
     /**
      * Default Constructor.
      */
@@ -152,29 +147,6 @@ public class CommonsPlugin extends Plugin {
     @Override
     public void start(BundleContext context) throws Exception {
         this.context = context;
-        configure();
-    }
-
-    private void configure() {
-        
-        try {
-          Properties props = new Properties();
-          this.logManager = new PerformanceLogManager(this, props);
-        } 
-        catch (Exception e) {
-          String message = "Error while initializing log properties." + 
-                           e.getMessage();
-          IStatus status = new Status(IStatus.ERROR,
-                  getDefault().getBundle().getSymbolicName(),
-                  IStatus.ERROR, message, e);
-          getLog().log(status);
-          throw new RuntimeException(
-               "Error while initializing log properties.",e);
-        }         
-      }
-    
-    public static PerformanceLogManager getLogManager() {
-        return getDefault().logManager; 
     }
 
     /*
@@ -184,10 +156,6 @@ public class CommonsPlugin extends Plugin {
      */
     public void stop(BundleContext context) throws Exception {
         plugin = null;
-        if (this.logManager != null) {
-            this.logManager.shutdown();
-            this.logManager = null;
-        }   
         super.stop(context);
     }
